gpu-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
ultralytics==8.3.74
opencv-python==4.10.0.84
numpy==2.1.3
pytile==2024.9.0
aiohttp==3.11.11
python-dotenv==1.0.1
